validate/requirements.txt
appyter @ git+https://github.com/Maayanlab/appyter
click
fsspec
jsonschema
nbformat
Pillow
requests